pyproject.toml
[project]
name = "deltachat-mcp"
version = "0.2.0"
description = "Model Context Protocol (MCP) server for Delta Chat — let AI send/receive encrypted messages"
authors = [{ name = "Your Name", email = "you@example.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "deltachat2>=0.6.0",
    "mcp>=1.19.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "black", "ruff"]

[project.scripts]
deltachat-mcp = "deltachat_mcp.server:main"

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
